Splitted and Reorganizes code
