+++ src/css/GameBoard.css
@@ -1,0 +1,100 @@
+.game-board {
+    text-align: center;
+    font-family: 'Poppins', sans-serif;
+    background: linear-gradient(135deg, #c73f93, #bdc00e);
+    width: 100vw;
+    height: 100vh;
+    color: #fff;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    margin: 0;
+    overflow: hidden;
+  }
+  
+  .game-area {
+    position: relative;
+    width: 80vw;
+    height: 80vh;
+    max-width: 800px;
+    max-height: 800px;
+    border: 5px solid #fff;
+    margin: 20px auto;
+    overflow: hidden;
+    background: linear-gradient(135deg, #00ffff, #09ee68);
+    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.25);
+    border-radius: 10px;
+  }
+  
+  .ball {
+    position: absolute;
+    width: 4vw;
+    height: 4vw;
+    max-width: 30px;
+    max-height: 30px;
+    background: radial-gradient(circle, #d15f63, #bb9186);
+    border: 2px solid #fff;
+    border-radius: 50%;
+    cursor: pointer;
+    animation: pulse 1s infinite;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.25);
+    transition: transform 0.2s ease;
+  }
+  
+  .ball:hover {
+    transform: scale(1.2);
+  }
+  
+  button {
+    padding: 1rem 2rem;
+    font-size: 1.2rem;
+    font-weight: bold;
+    color: #fff;
+    background: linear-gradient(135deg, #ff7e5f, #feb47b);
+    border: none;
+    border-radius: 50px;
+    cursor: pointer;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.25);
+    transition: transform 0.2s ease, box-shadow 0.2s ease;
+    margin: 20px;
+  }
+  
+  button:hover {
+    transform: scale(1.1);
+    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.35);
+  }
+  
+  /* Message Styling */
+.message {
+    position: absolute;
+    left: 50%;
+    transform: translateX(-50%);
+    font-size: 18px;
+    font-weight: bold;
+    color: #fff;
+    background: rgba(0, 0, 0, 0.5);
+    padding: 5px 15px;
+    border-radius: 10px;
+    pointer-events: none; 
+    z-index: 10;
+    animation: fadeOut 2s forwards; 
+  }
+  
+  @keyframes fadeOut {
+    0% {
+      opacity: 1;
+    }
+    100% {
+      opacity: 0;
+    }
+  }
+  
+  @keyframes fadeOut {
+    0% {
+      opacity: 1;
+    }
+    100% {
+      opacity: 0;
+    }
+  }
