+++ src/types/Ball.ts
@@ -1,0 +1,6 @@
+export interface Ball {
+    id: number;
+    x: number;
+    y: number;
+  }
+
